+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## Latest changed
+## Latest
 
+## CARLA-ROS-Bridge 0.9.12
+
+*   Fixed scenario runner node shutdown for foxy
+*   Fixed actor synchronization creation
 *   Fix of scenario runner status
 
 ## CARLA-ROS-Bridge 0.9.11